app/requirements.txt
great-expectations
textblob
boto3
newsapi-python
